refactor(channel): extract command orchestration foundation
