+++ crates/app/src/channel/commands/mod.rs
@@ -1,0 +1,7 @@
+pub(in crate::channel) mod context;
+mod send;
+mod serve;
+
+pub(super) use context::{ChannelCommandContext, ChannelResolvedRuntimeAccount};
+pub(super) use send::{ChannelSendCommandSpec, run_channel_send_command};
+pub(super) use serve::run_channel_serve_command_with_stop;
